Significant code refactoring
Changed the environment.yml file to allow users of different platforms to create an environement. Refactored the actor class to allow for different strategies of training (with or without memory and with or without a target model). Implemented a sliding window function for plotting in the Experiment file, alongside additional experiments to perform.